Genetic Algorithm version 1.0
ok.... so it doesn't do much atm except make a random string for genetic
info. for a population of 10 and breed 10 alpha organism focused
generations but its setup so we can eventually add more complex "genetic
info" i.e. the solution to the problem the AI is trying to solve